+++ packages/design-system-react-native/src/components/TextButton/index.ts
@@ -1,3 +1,2 @@
-export { TextButtonSize } from '../../types';
 export { TextButton } from './TextButton';
 export type { TextButtonProps } from './TextButton.types';