+++ packages/design-system-react/src/components/Checkbox/Checkbox.tsx
@@ -66,7 +66,7 @@ export const Checkbox = forwardRef<{ toggle: () => void }, CheckboxProps>(
     }
 
     const checkboxClasses = twMerge(
-      'relative flex size-6 items-center justify-center rounded border-2 p-0 transition-transform active:scale-95',
+      'relative flex size-6 items-center justify-center rounded border-2 p-0 transition-transform active:scale-95 peer-focus-visible:outline peer-focus-visible:outline-2 peer-focus-visible:outline-offset-2 peer-focus-visible:outline-primary-default',
       baseBg,
       baseBorder,
       checkboxContainerProps?.className,
@@ -89,7 +89,7 @@ export const Checkbox = forwardRef<{ toggle: () => void }, CheckboxProps>(
             aria-invalid={isInvalid}
             onChange={handleChange}
             onKeyDown={handleKeyDown}
-            className="absolute inset-0 size-full cursor-pointer opacity-0 disabled:cursor-not-allowed"
+            className="peer absolute inset-0 size-full cursor-pointer opacity-0 disabled:cursor-not-allowed"
             {...inputProps}
           />
           <div className={checkboxClasses} {...checkboxContainerProps}>